refactor(test-cases): split the test_cases CLI into a subpackage
test_cases.py had grown past 1700 lines. Split it into a
climate_ref.cli.test_cases package grouped by concern: discovery
(fetch/list), run, baselines (replay/mint/build), store
(sync/check-store) and ci_gate, over _catalog/_common/_stages helpers
and an _app module holding the shared Typer app.

Command implementations are carried over unchanged. The package
re-exports the helpers the unit tests import directly and registers the
command modules in a fixed order so `ref test-cases --help` keeps its
verb ordering. _test_case_stages.py moves in as _stages.py; the test
mock targets are updated to the new module paths.
